+++ ase/io/vasp.py
@@ -327,7 +327,7 @@ def read_vasp_xdatcar(filename='XDATCAR', index=-1):
         return images[index]
 
 
-def __get_xml_parameter(par):
+def __get_xml_parameter(par):  # noqa: N802
     """An auxiliary function that enables convenient extraction of
     parameter values from a vasprun.xml file with proper type
     handling.
@@ -385,13 +385,20 @@ def read_vasp_xml(filename='vasprun.xml', index=-1):
 
             if event == 'end':
                 if elem.tag == 'kpoints':
-                    for subelem in elem.iter(tag='generation'):
-                        kpts_params = OrderedDict()
-                        parameters['kpoints_generation'] = kpts_params
-                        for par in subelem.iter():
-                            if par.tag in ['v', 'i']:
-                                parname = par.attrib['name'].lower()
-                                kpts_params[parname] = __get_xml_parameter(par)
+                    generation_block = list(elem.iter(tag='generation'))
+                    # This block can also be 'listgenerated', in which case we
+                    # don't try to read it: the format is different.
+                    if (generation_block
+                        and generation_block[0]
+                            .attrib['param'] == 'Monkhorst-Pack'):
+                        for subelem in generation_block:
+                            kpts_params = OrderedDict()
+                            parameters['kpoints_generation'] = kpts_params
+                            for par in subelem.iter():
+                                if par.tag in ['v', 'i']:
+                                    parname = par.attrib['name'].lower()
+                                    kpts_params[parname] = \
+                                      __get_xml_parameter(par)
 
                     kpts = elem.findall("varray[@name='kpointlist']/v")
                     ibz_kpts = np.zeros((len(kpts), 3))
@@ -485,6 +492,7 @@ def read_vasp_xml(filename='vasprun.xml', index=-1):
         # e_0_energy - e_fr_energy from calculation/scstep/energy, then
         # apply that correction to e_fr_energy from calculation/energy.
         lastscf = step.findall('scstep/energy')[-1]
+
         dipoles = step.findall('scstep/dipole')
         if dipoles:
             lastdipole = dipoles[-1]